backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
motor==3.3.2
pymongo==4.6.0

# AI/LLM
openai==1.3.8

# Google Sheets
gspread==5.11.3
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# HTTP & Async
httpx==0.25.2
aiohttp==3.9.1

# Logging & Monitoring
python-json-logger==2.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Security
python-multipart==0.0.6
